requirements.txt
langchain-openai>=0.0.5
python-dotenv>=1.0.0
browser-use>=0.1.0
openai>=1.12.0
playwright>=1.42.0
pillow>=10.2.0
aiofiles>=23.2.1
websockets>=12.0
asyncio>=3.4.3
nest-asyncio>=1.6.0
tqdm>=4.66.2
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0 